Rename TestProjectContext files
