+++ Tests/BabelWiresLib/changeFileCommandTest.cpp
@@ -12,7 +12,7 @@
 
 #include "Tests/BabelWiresLib/TestUtils/testFeatureElement.hpp"
 #include "Tests/BabelWiresLib/TestUtils/testFileFormats.hpp"
-#include "Tests/BabelWiresLib/TestUtils/testProjectContext.hpp"
+#include "Tests/BabelWiresLib/TestUtils/testEnvironment.hpp"
 #include "Tests/TestUtils/tempFilePath.hpp"
 
 namespace {
