+++ Tests/BabelWiresLib/featureElementConnectionTest.cpp
@@ -8,7 +8,7 @@
 #include "BabelWiresLib/Project/Modifiers/connectionModifierData.hpp"
 
 #include "Tests/BabelWiresLib/TestUtils/testFeatureElement.hpp"
-#include "Tests/BabelWiresLib/TestUtils/testProjectContext.hpp"
+#include "Tests/BabelWiresLib/TestUtils/testEnvironment.hpp"
 #include "Tests/BabelWiresLib/TestUtils/testRecord.hpp"
 #include "Tests/TestUtils/testLog.hpp"
