+++ Tests/BabelWiresLib/processorElementTest.cpp
@@ -9,7 +9,7 @@
 #include "Common/Identifiers/identifierRegistry.hpp"
 
 #include "Tests/BabelWiresLib/TestUtils/testProcessor.hpp"
-#include "Tests/BabelWiresLib/TestUtils/testProjectContext.hpp"
+#include "Tests/BabelWiresLib/TestUtils/testEnvironment.hpp"
 #include "Tests/BabelWiresLib/TestUtils/testRecord.hpp"
 
 #include "Tests/TestUtils/tempFilePath.hpp"
