+++ Tests/BabelWiresLib/projectSerializationTest.cpp
@@ -4,7 +4,7 @@
 #include "BabelWiresLib/Serialization/projectSerialization.hpp"
 
 #include "Tests/BabelWiresLib/TestUtils/testFileFormats.hpp"
-#include "Tests/BabelWiresLib/TestUtils/testProjectContext.hpp"
+#include "Tests/BabelWiresLib/TestUtils/testEnvironment.hpp"
 #include "Tests/BabelWiresLib/TestUtils/testProjectData.hpp"
 #include "Tests/BabelWiresLib/TestUtils/testRecord.hpp"
 #include "Tests/TestUtils/tempFilePath.hpp"
