+++ Tests/BabelWiresLib/targetFileElementTest.cpp
@@ -8,7 +8,7 @@
 #include "BabelWiresLib/Features/numericFeature.hpp"
 
 #include "Tests/BabelWiresLib/TestUtils/testFileFormats.hpp"
-#include "Tests/BabelWiresLib/TestUtils/testProjectContext.hpp"
+#include "Tests/BabelWiresLib/TestUtils/testEnvironment.hpp"
 
 #include "Tests/TestUtils/tempFilePath.hpp"
